Add component to protected routes private
ProtectedRoute check if user is one of three states:

Initial, which is an empty object {}, in that case render an empty div.

Not logged in, which is null, in this case redirect to home page "/".

Has logged in, which is a non-empty object. In this case render the requested route.