Add batch firestore function to add post data when user creates a post
Create new post doc in posts collection.

Set the post-like-list doc of selected post in post-like-list collection. This document will be used to save the list of users who have liked this post.

Update the count doc with the selected post in post-like-list collection. This document will be a summary with the key value of postId and number of likes, of all posts.

This batch function ensures that all data related to the post is always created when the operation is successful.